Corrected type hint

+++ Area_Calculator.py
@@ -70,8 +70,8 @@ def find_peak_near_extremes(
 def analyze_data(
     x_input: np.ndarray[np.uint16],
     y_input: np.ndarray[np.uint16],
-    tablet_width_mm: int,
-    tablet_height_mm: int,
+    tablet_width_mm: float,
+    tablet_height_mm: float,
     innergameplay_width_px: int,
     innergameplay_height_px: int,
 ):
@@ -136,6 +136,7 @@ def main(
         int, typer.Option(prompt="Enter map duration in seconds", min=10)
     ],
 ):
+
     innergameplay_height_px = int((864 / 1080) * screen_height_px)
     innergameplay_width_px = int((1152 / 1920) * screen_width_px)
     typer.confirm(
